HomeWork2  refactoring
